Added static casts to prevend various warnings.

+++ Engine.h
@@ -24,7 +24,7 @@ private:
 	void afterEffects();
 	void addBloom();
 	void addRenderLines();
-	sf::VertexArray createRenderLine(const sf::Vector2f& start, const sf::Vector2f& end, float thickness, sf::Color color);
+	sf::VertexArray createRenderLine(const sf::Vector2f& start, const sf::Vector2f& end, int thickness, sf::Color color);
 
 	sf::Clock m_clock;
 	sf::Time m_elapsed;
